[BST-21] feat: Board mapper 기능 구현
